requirements.txt
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
pillow==11.3.0
psycopg2-binary==2.9.10
pytest==8.4.2
pytest-django==4.11.1